requirements.txt
mysql-connector-python
pyTelegramBotAPI
virtualenv
requests
python-dotenv